api/requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
pydantic==2.5.3
PyGithub==2.1.1
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0 